Refactor authentication models and services
Introduced the `NSE.Shared` namespace and added new classes
(`UserClaim`, `UserLoginResponse`, `UserToken`) for
authentication and user management. Updated `AuthProvider`
and `IAuthProvider` to reference the new models. Added
`AuthHttpIntegrationService` and `IAuthHttpIntegrationService`
for handling user registration and sign-in. Updated
`NSE.WebApp.MVC.csproj` to include a reference to the
`NSE.Shared` project.